requirements.txt
line-bot-sdk==3.5.0
Flask==3.1.1
openai==1.97.1
pandas
opencv-python
scikit-learn
python-dotenv
